+++ lib/http_stub/server/request/request.rb
@@ -4,16 +4,16 @@ module HttpStub
 
       class Request
 
-        attr_reader :base_uri, :uri, :method, :headers, :parameters, :body, :session
+        attr_reader :base_uri, :uri, :method, :headers, :parameters, :body
+        attr_accessor :session
 
-        def initialize(rack_request, session)
+        def initialize(rack_request)
           @base_uri   = rack_request.base_url
           @uri        = rack_request.path_info
           @method     = rack_request.request_method.downcase
           @headers    = HttpStub::Server::Request::Headers.create(rack_request)
           @parameters = HttpStub::Server::Request::Parameters.create(rack_request)
           @body       = rack_request.body.read
-          @session    = session
         end
 
         def to_json(*args)