+++ spec/lib/http_stub/rake/server_tasks_spec.rb
@@ -9,25 +9,30 @@ describe HttpStub::Rake::ServerTasks do
 
   end
 
-  let(:server_name)  { "server_tasks_test_#{HttpStub::Rake::ServerTasksTestCounter.next}".to_sym }
-  let(:default_args) { { port: 8001, name: server_name } }
-  let(:args)         { {} }
-  let(:configurer)   { double(HttpStub::Configurer).as_null_object }
+  let(:server_name)     { "server_tasks_test_#{HttpStub::Rake::ServerTasksTestCounter.next}".to_sym }
+  let(:default_args)    { { port: 8001, name: server_name } }
+  let(:additional_args) { {} }
+  let(:args)            { default_args.merge(additional_args) }
+  let(:configurer)      { double(HttpStub::Configurer).as_null_object }
 
-  before(:example) { HttpStub::Rake::ServerTasks.new(default_args.merge(args)) }
+  before(:example) { HttpStub::Rake::ServerTasks.new(args) }
 
   describe "the configure task" do
 
-    context "when a configurer is provided" do
+    let(:task) { Rake::Task["#{server_name}:configure"] }
 
-      let(:args) { { configurer: configurer } }
+    context "when executed" do
 
-      context "and the task is executed" do
+      subject { task.execute }
+
+      context "with a configurer" do
+
+        let(:additional_args) { { configurer: configurer } }
 
         it "initializes the provided configurer" do
           expect(configurer).to receive(:initialize!)
 
-          Rake::Task["#{server_name}:configure"].execute
+          subject
         end
 
       end
@@ -37,7 +42,39 @@ describe HttpStub::Rake::ServerTasks do
     context "when a configurer is not provided" do
 
       it "does not generate a task" do
-        expect { Rake::Task["#{server_name}:configure"] }.to raise_error(/Don't know how to build task/)
+        expect { task }.to raise_error(/Don't know how to build task/)
+      end
+
+    end
+
+  end
+
+  describe "the start foreground task" do
+
+    let(:application) { HttpStub::Server::Application::Application }
+
+    let(:task) { Rake::Task["#{server_name}:start:foreground"] }
+
+    context "when executed" do
+
+      subject { task.execute }
+
+      before(:example) do
+        allow(HttpStub::Server::Application::Application).to receive(:configure)
+        allow(HttpStub::Server::Application::Application).to receive(:run!)
+      end
+
+      it "configures the application using the tasks arguments" do
+        expect(application).to receive(:configure).with(args)
+
+        subject
+      end
+
+      it "runs the configured application" do
+        expect(application).to receive(:configure).ordered
+        expect(application).to receive(:run!).ordered
+
+        subject
       end
 
     end