+++ spec/lib/http_stub/server/request/factory_spec.rb
@@ -1,43 +1,57 @@
 describe HttpStub::Server::Request::Factory do
 
-  let(:scenario_registry) { instance_double(HttpStub::Server::Registry) }
+  let(:scenario_registry)     { instance_double(HttpStub::Server::Registry) }
+  let(:session_configuration) { { some_configuration_key: "some configuration value" } }
 
   let(:rack_request) { Rack::RequestFixture.create }
 
-  let(:factory) { described_class.new(scenario_registry) }
+  let(:factory) { described_class.new(scenario_registry, session_configuration) }
 
-  describe "::create" do
+  describe "constructor" do
 
-    let(:session) { instance_double(HttpStub::Server::Session) }
+
+
+
+  end
+
+  describe "#create" do
+
+    let(:session_factory) { instance_double(HttpStub::Server::Session::Factory, create: session) }
+    let(:session)         { instance_double(HttpStub::Server::Session::Session) }
+
+    let(:created_request) { instance_double(HttpStub::Server::Request::Request, "session=" => nil) }
 
     subject { factory.create(rack_request) }
 
+    before(:example) { allow(HttpStub::Server::Session::Factory).to receive(:new).and_return(session_factory) }
+
     it "creates a request for the rack request" do
-      expect(HttpStub::Server::Request::Request).to receive(:new).with(rack_request, anything)
+      expect(HttpStub::Server::Request::Request).to receive(:new).with(rack_request).and_return(created_request)
 
       subject
     end
 
-    it "creates a session" do
-      expect(HttpStub::Server::Session).to receive(:new).with(scenario_registry)
+    it "creates a session factory using the factories scenario registry and configuration" do
+      expect(HttpStub::Server::Session::Factory).to receive(:new).with(scenario_registry, session_configuration)
 
       subject
     end
 
-    it "creates a session that is shared across all requests" do
-      allow(HttpStub::Server::Session).to receive(:new).once.and_return(session)
-      rack_requests = (1..3).map { Rack::RequestFixture.create }
+    it "creates a session for the request via the session factory" do
+      allow(HttpStub::Server::Request::Request).to receive(:new).and_return(created_request)
+      expect(session_factory).to receive(:create).with(created_request)
 
-      requests = rack_requests.map { |rack_request| factory.create(rack_request) }
+      subject
+    end
 
-      requests.each { |request| expect(request.session).to eql(session) }
+    it "establishes the session on the request" do
+      expect(subject.session).to eql(session)
     end
 
     it "returns the created request" do
-      http_stub_request = instance_double(HttpStub::Server::Request::Request)
-      allow(HttpStub::Server::Request::Request).to receive(:new).and_return(http_stub_request)
+      allow(HttpStub::Server::Request::Request).to receive(:new).and_return(created_request)
 
-      expect(subject).to eql(http_stub_request)
+      expect(subject).to eql(created_request)
     end
 
   end